+++ src/features/basic/browser-tab-name.ts
